ai_agent_backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-RESTful==0.3.9
Flask-CORS==3.0.10
Flask-SQLAlchemy==2.5.1
PyJWT==2.3.0
SQLAlchemy==1.4.23